directory cleanup
